+++ packages/cli/examples/umd/iwr.config.js
@@ -1,5 +1,0 @@
-module.exports = {
-  analyzerPort: 8888,
-  target: "./src",
-  buildPath: "./dist",
-};