+++ packages/cli/examples/umd/src/react/index.ts
@@ -1,2 +1,0 @@
-export { default as BReact } from "react";
-export { default as BReactDom } from "react-dom";