+++ packages/template/packages/react/src/app.jsx
@@ -1,5 +1,4 @@
 import './index.scss';
-import './assets/images/favicon.ico';
 
 import React from 'react';
 import { render } from 'react-dom';